reorganized, started moving features out of bot.js
