+++ src/controller/commands/roleCommand.js
@@ -1,5 +1,5 @@
-const message = require('../functions/message');
-const Command = require('../models/Command');
+const message = require('../../functions/message');
+const Command = require('../../models/Command');
 
 const changeColor = (msg, rest, roleName) => {
     const color = rest[1];
